feat(services/ipfs): implement ipfs_stat
